# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admrentas API",
        "description": "Documentacion de APIs de Admrentas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{scheme}://{host}",
            "description": "Servidor actual por entorno",
            "variables": {
                "scheme": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                },
                "host": {
                    "default": "localhost"
                }
            }
        }
    ],
    "paths": {
        "/api/advisors/suggestions": {
            "get": {
                "tags": [
                    "Advisors"
                ],
                "summary": "Sugerencias de estados, ciudades y asesores",
                "operationId": "9a540aa54e44f2146795b4ed8c192d8b",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sugerencias obtenidas"
                    }
                }
            }
        },
        "/api/advisors/search": {
            "get": {
                "tags": [
                    "Advisors"
                ],
                "summary": "Buscar asesores",
                "operationId": "68d821642128baad7cab4d2d4f098bb9",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estate_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados paginados"
                    },
                    "422": {
                        "description": "Parámetros inválidos"
                    }
                }
            }
        },
        "/api/advisors/details": {
            "get": {
                "tags": [
                    "Advisors"
                ],
                "summary": "Detalles de asesores por criterio de coincidencia",
                "operationId": "353408e98668ece1da8c6a12e012b240",
                "parameters": [
                    {
                        "name": "matched_by",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "state",
                                "city"
                            ]
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados paginados"
                    },
                    "422": {
                        "description": "Parámetros inválidos"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}